+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix cronitor.io integration in proxy environments.
+
 ## [0.48.0] - 2025-11-04
 
 ### Added